# frontend/openapi.json
{
  "components" : {
    "schemas" : {
...
      "CurrencySummary" : {
        "properties" : {
          "byGameType" : {
            "description" : "The game types with finished games, in catalog order",
            "items" : {
              "$ref" : "#/components/schemas/GameTypeSummary"
            },
            "type" : "array"
          },
          "currencyCode" : {
            "type" : "string"
          },
          "inPlay" : {
            "$ref" : "#/components/schemas/InPlay"
          },
          "total" : {
            "$ref" : "#/components/schemas/StatsFigures",
            "description" : "Every finished game in this currency"
          }
        },
        "required" : [ "byGameType", "currencyCode", "inPlay", "total" ],
        "type" : "object"
      },
...
      "GameTypeSummary" : {
        "properties" : {
          "figures" : {
            "$ref" : "#/components/schemas/StatsFigures"
          },
          "gameType" : {
            "enum" : [ "TOURNAMENT", "SIT_AND_GO", "CASH" ],
            "type" : "string"
          }
        },
        "required" : [ "figures", "gameType" ],
        "type" : "object"
      },
...
      "InPlay" : {
        "properties" : {
          "games" : {
            "format" : "int64",
            "type" : "integer"
          },
          "invested" : {
            "description" : "Money paid so far for their entries",
            "type" : "number"
          }
        },
        "required" : [ "games", "invested" ],
        "type" : "object"
      },
      "MovementRequest" : {
        "properties" : {
          "amount" : {
            "description" : "Greater than zero: the type gives the direction (a withdrawal subtracts). Only an ADJUSTMENT can be negative",
            "type" : "number"
          },
          "currencyCode" : {
            "description" : "Only for a movement without a room, e.g. EUR",
            "type" : [ "string", "null" ]
          },
          "notes" : {
            "maxLength" : 5000,
            "minLength" : 0,
            "type" : [ "string", "null" ]
          },
          "occurredOn" : {
            "format" : "date",
            "type" : "string"
          },
          "roomId" : {
            "description" : "Room the movement belongs to; the amount is in its currency. Omit it, and give a currency, for a movement of the bankroll as a whole",
            "format" : "int64",
            "type" : [ "integer", "null" ]
          },
          "type" : {
            "enum" : [ "DEPOSIT", "WITHDRAWAL", "BONUS", "ADJUSTMENT" ],
            "type" : "string"
          }
        },
        "required" : [ "amount", "occurredOn", "type" ],
        "type" : "object"
      },
      "MovementResponse" : {
        "properties" : {
          "amount" : {
            "type" : "number"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "currencyCode" : {
            "description" : "Currency of the amount: the one of the room, or its own without a room",
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "notes" : {
            "type" : [ "string", "null" ]
          },
          "occurredOn" : {
            "format" : "date",
            "type" : "string"
          },
          "room" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/RoomRef"
            }, {
              "type" : "null"
            } ],
            "description" : "Null for a movement of the bankroll as a whole"
          },
          "signedAmount" : {
            "description" : "Effect on the bankroll: the amount, negative for a withdrawal",
            "type" : "number"
          },
          "type" : {
            "enum" : [ "DEPOSIT", "WITHDRAWAL", "BONUS", "ADJUSTMENT" ],
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "required" : [ "amount", "createdAt", "currencyCode", "id", "occurredOn", "signedAmount", "type", "updatedAt" ],
        "type" : "object"
      },
...
      "PageResponseMovementResponse" : {
        "properties" : {
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/MovementResponse"
            },
            "type" : "array"
          },
          "page" : {
            "description" : "Zero-based page number",
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "description" : "Requested page size",
            "format" : "int32",
            "type" : "integer"
          },
          "totalItems" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "items", "page", "size", "totalItems", "totalPages" ],
        "type" : "object"
      },
...
      "RoomRef" : {
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          }
        },
        "required" : [ "id", "name" ],
        "type" : "object"
      },
...
      "StatsFigures" : {
        "properties" : {
          "averageBuyIn" : {
            "description" : "Mean buy-in of a game; cash games apart",
            "type" : [ "number", "null" ]
          },
          "bounties" : {
            "description" : "Bounties, already included in won",
            "type" : "number"
          },
          "entries" : {
            "description" : "Entries, re-entries included",
            "format" : "int64",
            "type" : "integer"
          },
          "games" : {
            "description" : "Games played",
            "format" : "int64",
            "type" : "integer"
          },
          "gamesInTheMoney" : {
            "description" : "Games with a prize or a ticket won (bounties do not count); cash games apart",
            "format" : "int64",
            "type" : [ "integer", "null" ]
          },
          "gamesWithPrize" : {
            "description" : "Games with a prize, a bounty or a ticket won; cash games apart",
            "format" : "int64",
            "type" : [ "integer", "null" ]
          },
          "inTheMoneyRate" : {
            "description" : "gamesInTheMoney / games, cash games apart",
            "type" : [ "number", "null" ]
          },
          "invested" : {
            "description" : "Money paid for the entries (an entry paid with a ticket costs nothing)",
            "type" : "number"
          },
          "net" : {
            "description" : "won - invested",
            "type" : "number"
          },
          "roi" : {
            "description" : "net / invested; null when nothing was invested",
            "type" : [ "number", "null" ]
          },
          "ticketsWon" : {
            "description" : "Value of the tickets won; informative, not part of won, net or roi",
            "type" : "number"
          },
          "winningGames" : {
            "description" : "Games with a positive net",
            "format" : "int64",
            "type" : "integer"
          },
          "winningRate" : {
            "description" : "winningGames / games",
            "type" : [ "number", "null" ]
          },
          "withPrizeRate" : {
            "description" : "gamesWithPrize / games, cash games apart",
            "type" : [ "number", "null" ]
          },
          "won" : {
            "description" : "Money won: prizes + bounties",
            "type" : "number"
          }
        },
        "required" : [ "bounties", "entries", "games", "invested", "net", "ticketsWon", "winningGames", "won" ],
        "type" : "object"
      },
...
      "StatsSummaryResponse" : {
        "properties" : {
          "currencies" : {
            "items" : {
              "$ref" : "#/components/schemas/CurrencySummary"
            },
            "type" : "array"
          }
        },
        "required" : [ "currencies" ],
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "description" : "REST API of poker-bankroll, a self-hosted poker bankroll manager.",
    "license" : {
      "name" : "MIT",
      "url" : "https://opensource.org/licenses/MIT"
    },
    "title" : "poker-bankroll API"
  },
  "openapi" : "3.1.0",
  "paths" : {
    "/bankroll/movements" : {
      "get" : {
        "description" : "All filters are optional and combined with AND.",
        "operationId" : "list_3",
        "parameters" : [ {
          "description" : "On or after this date",
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        }, {
          "description" : "On or before this date",
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "type",
          "required" : false,
          "schema" : {
            "enum" : [ "DEPOSIT", "WITHDRAWAL", "BONUS", "ADJUSTMENT" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "roomId",
          "required" : false,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "description" : "true: only movements that belong to no room; false: only those of a room",
          "in" : "query",
          "name" : "withoutRoom",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "description" : "Currency of the amount, e.g. EUR",
          "in" : "query",
          "name" : "currency",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Zero-based page number",
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "minimum" : 0,
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 50,
            "format" : "int32",
            "maximum" : 200,
            "minimum" : 1,
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseMovementResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "List bankroll movements, newest first",
        "tags" : [ "Bankroll" ]
      },
      "post" : {
        "description" : "Of a room (`roomId`), or of the bankroll as a whole (`currencyCode`, no room).",
        "operationId" : "create_3",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MovementRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MovementResponse"
                }
              }
            },
            "description" : "Created"
          }
        },
        "summary" : "Record a bankroll movement",
        "tags" : [ "Bankroll" ]
      }
    },
    "/bankroll/movements/{id}" : {
      "delete" : {
        "operationId" : "delete_3",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "summary" : "Delete a bankroll movement",
        "tags" : [ "Bankroll" ]
      },
      "get" : {
        "operationId" : "get_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MovementResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Get a bankroll movement",
        "tags" : [ "Bankroll" ]
      },
      "put" : {
        "description" : "Replaces every field.",
        "operationId" : "update_3",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MovementRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MovementResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Update a bankroll movement",
        "tags" : [ "Bankroll" ]
      }
    },
...
    "/stats/summary" : {
      "get" : {
        "description" : "Takes the filters of the games list, all optional and combined with AND. Only finished games count in the figures; games in play are reported apart.",
        "operationId" : "summary",
        "parameters" : [ {
          "description" : "Played on or after this date",
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        }, {
          "description" : "Played on or before this date",
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "gameType",
          "required" : false,
          "schema" : {
            "enum" : [ "TOURNAMENT", "SIT_AND_GO", "CASH" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "modality",
          "required" : false,
          "schema" : {
            "enum" : [ "NLHE", "PLO" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "roomId",
          "required" : false,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "variantId",
          "required" : false,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "description" : "Currency of the room, e.g. EUR",
          "in" : "query",
          "name" : "currency",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Text contained in the name or the notes, ignoring case",
          "in" : "query",
          "name" : "q",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StatsSummaryResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Results overall and per game type, for each currency",
        "tags" : [ "Statistics" ]
      }
    },
...
  },
...
}